Phuoc: push some update for module product

+++ resources/views/admin/products/view.blade.php
@@ -14,8 +14,10 @@
         var sum = 0;
         var html = '';
         $.each(arr_details, function(key, value){
-            html += '<tr class="child" data-index="'+key+'"><td><a href="javascript:;" onclick="deleteProductInfoItem('+key+');">Delete</a></td><td>'+value['color']['name']+'</td><td>'+value['size']['name']+'</td><td class="c-quantity">'+value['quantity']+'</td></tr>';
-            sum += parseInt(value['quantity']);
+            if (value.delete != true) {
+                html += '<tr class="child" data-index="'+key+'"><td><a href="javascript:;" onclick="deleteProductInfoItem('+key+');">Delete</a></td><td>'+value['color']['name']+'</td><td>'+value['size']['name']+'</td><td class="c-quantity">'+value['quantity']+'</td></tr>';
+                sum += parseInt(value['quantity']);
+            }
         });
         html += '<tr><td></td><td></td><td></td><td>Tổng số lượng: <span class="c-total-quantities">'+sum+'</span></td></tr>';
         $('#i-product-info tbody').html(html);
@@ -26,7 +28,8 @@
     }
 
     function deleteProductInfoItem(key) {
-        details.splice( $.inArray(key, details), 1 );
+        details[key].delete = true;
+        // details.splice( $.inArray(key, details), 1 );
         print_table_details(details);
     }
 
@@ -137,7 +140,11 @@
         print_table_photos(photos);
     }
 
-    $(document).ready(function () {
+    function synchronize_child_and_parent_category($) {
+
+    }
+
+    $(document).ready(function ($) {
         $( "#mainForm" ).submit(function( event ) {
             var searchIDs = $("#mainForm .list-tree-section input:checkbox:checked").map(function(){
               return $(this).val();
@@ -156,6 +163,58 @@
 
         print_table_details(details);
         print_table_photos(photos);
+        synchronize_child_and_parent_category($);
+
+
+        $('input[type="checkbox"]').change(function(e) {
+
+          var checked = $(this).prop("checked"),
+          container = $(this).parent(),
+          siblings = container.siblings();
+
+          container.find('input[type="checkbox"]').prop({
+            indeterminate: false,
+            checked: checked
+        });
+
+          function checkSiblings(el) {
+
+            var parent = el.parent().parent(),
+            all = true;
+
+            el.siblings().each(function() {
+              return all = ($(this).children('input[type="checkbox"]').prop("checked") === checked);
+          });
+
+            if (all && checked) {
+
+              parent.children('input[type="checkbox"]').prop({
+                indeterminate: false,
+                checked: checked
+            });
+
+              checkSiblings(parent);
+
+          } else if (all && !checked) {
+
+              parent.children('input[type="checkbox"]').prop("checked", checked);
+              parent.children('input[type="checkbox"]').prop("indeterminate", (parent.find('input[type="checkbox"]:checked').length > 0));
+              checkSiblings(parent);
+
+          } else {
+
+              el.parents("li").children('input[type="checkbox"]').prop({
+            // indeterminate: true,
+            checked: false
+        });
+
+          }
+
+      }
+
+      checkSiblings(container);
+  });
+        
 
     });
 </script>
